remove tui.go and migrate code into config.go
